refactor(native): move chunk stream module and tests into chunk folders
